ver 0.008 -fixed bug for reshuffling big data. -fixed some comments (grammatic errors)
git-svn-id: svn+ssh://svn.r-forge.r-project.org/svnroot/genabel/pkg/OmicABEL@1260 edb9625f-4e0d-4859-8d74-9fd3b1da38cb

+++ src/reshuffle/iout_file.cpp
@@ -78,8 +78,8 @@ ostream &operator <<(ostream &os,labels_data labels) {
 	os <<"NUMBER OF TRAITNAMES\t"<<(*labels.trait_names).size()<<endl;
 }
 
-int iout_file::tilecoordinates(int traitNo, int snpNo) {
-	int tileCoor = 0;
+int64_t iout_file::tilecoordinates(int traitNo, int snpNo) {
+	int64_t tileCoor = 0;
 	int t_tile = traitNo / header.tile_t;
 	int t_off = traitNo % header.tile_t;
 	int m_tile = snpNo / header.tile_m;
