+++ src/reshuffle/iout_file.h
@@ -53,6 +53,6 @@ public:
 	iout_header header;
 	labels_data labels;
 	iout_file(Parameters &);
-	int tilecoordinates(int, int);
+	int64_t tilecoordinates(int, int);
 };
 #endif /* IOUT_FILE_H_ */
